+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--
+- `SafeErc20` Utility. #289
+- Finite Fields arithmetics. #376
 
 ### Changed